% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_population_data.R
\name{read_population_data}
\alias{read_population_data}
\title{Read population data}
\usage{
read_population_data(root)
}
\arguments{
\item{root}{Path to the simulation folder}
}
\value{
A tibble
}
\description{
This function is a wrapper around \code{read_data} tailored at reading the
output file "traitmean.dat" into a tibble with one row per
time point.
}
\examples{

root <- system.file("extdata", "sim-example", package = "hemigener")
read_trait_mean_data(root)

}
\seealso{
\code{read_data}
}
